+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+1. Implemented service lock for `load-acl` and `store-acl` commands.
+
+### Changed
 1. Moved default `git` branch to `main`, in line with current development practice.
+
 
 ## [0.8.2](https://github.com/uhppoted/uhppoted-app-s3/releases/tag/v0.8.2) - 2022-10-14